[BOOK-79] refactor: infra user impl 분리 (#15)
